requirements.txt
Django==3.2.7
torch==1.10.0
torchvision==0.11.1
Pillow==8.4.0
